Move global rendering into a module
