+++ bin/model/app.rs
@@ -8,7 +8,7 @@ use wgpu;
 
 pub struct ResourceView {
     model: model::RenderModel,
-    global: render::GlobalContext,
+    global: render::global::Context,
     object: render::object::Context,
     transform: space::Transform,
     cam: space::Camera,
@@ -30,7 +30,7 @@ impl ResourceView {
         let mut init_encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
             todo: 0,
         });
-        let global = render::GlobalContext::new(device);
+        let global = render::global::Context::new(device);
         let object = render::object::Context::new(&mut init_encoder, device, &pal_data, &global);
         device.get_queue().submit(&[
             init_encoder.finish(),
@@ -126,7 +126,7 @@ impl Application for ResourceView {
     ) -> Vec<wgpu::CommandBuffer> {
         let mut updater = render::Updater::new(device);
         updater.update(&self.global.uniform_buf, &[
-            render::GlobalConstants::new(&self.cam, &self.light_config),
+            render::global::Constants::new(&self.cam, &self.light_config),
         ]);
 
         let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
