+++ src/render/mod.rs
@@ -14,6 +14,7 @@ use std::mem;
 
 //mod collision; ../TODO
 mod debug;
+pub mod global;
 pub mod object;
 pub mod terrain;
 
@@ -43,30 +44,6 @@ pub struct ShapePolygon {
     pub indices: [u16; 4],
     pub normal: [i8; 4],
     pub origin_square: [f32; 4],
-}
-
-#[derive(Clone, Copy)]
-pub struct GlobalConstants {
-    _camera_pos: [f32; 4],
-    _m_vp: [[f32; 4]; 4],
-    _m_inv_vp: [[f32; 4]; 4],
-    _light_pos: [f32; 4],
-    _light_color: [f32; 4],
-}
-
-impl GlobalConstants {
-    pub fn new(cam: &Camera, light: &settings::Light) -> Self {
-        use cgmath::SquareMatrix;
-
-        let mx_vp = cam.get_view_proj();
-        GlobalConstants {
-            _camera_pos: cam.loc.extend(1.0).into(),
-            _m_vp: mx_vp.into(),
-            _m_inv_vp: mx_vp.invert().unwrap().into(),
-            _light_pos: light.pos,
-            _light_color: light.color,
-        }
-    }
 }
 
 pub struct Updater<'a> {
@@ -99,73 +76,6 @@ impl<'a> Updater<'a> {
 
     pub fn finish(self) -> wgpu::CommandBuffer {
         self.command_encoder.finish()
-    }
-}
-
-
-pub struct GlobalContext {
-    pub uniform_buf: wgpu::Buffer,
-    pub palette_sampler: wgpu::Sampler,
-    pub bind_group: wgpu::BindGroup,
-    pub bind_group_layout: wgpu::BindGroupLayout,
-}
-
-impl GlobalContext {
-    pub fn new(device: &wgpu::Device) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            bindings: &[
-                wgpu::BindGroupLayoutBinding {
-                    binding: 0,
-                    visibility: wgpu::ShaderStageFlags::VERTEX | wgpu::ShaderStageFlags::FRAGMENT,
-                    ty: wgpu::BindingType::UniformBuffer,
-                },
-                wgpu::BindGroupLayoutBinding { // palette sampler
-                    binding: 1,
-                    visibility: wgpu::ShaderStageFlags::FRAGMENT,
-                    ty: wgpu::BindingType::Sampler,
-                },
-            ],
-        });
-        let uniform_buf = device.create_buffer(&wgpu::BufferDescriptor {
-            size: mem::size_of::<GlobalConstants>() as u32,
-            usage: wgpu::BufferUsageFlags::UNIFORM | wgpu::BufferUsageFlags::TRANSFER_DST,
-        });
-        let palette_sampler = device.create_sampler(&wgpu::SamplerDescriptor {
-            r_address_mode: wgpu::AddressMode::ClampToEdge,
-            s_address_mode: wgpu::AddressMode::ClampToEdge,
-            t_address_mode: wgpu::AddressMode::ClampToEdge,
-            mag_filter: wgpu::FilterMode::Linear,
-            min_filter: wgpu::FilterMode::Nearest,
-            mipmap_filter: wgpu::FilterMode::Nearest,
-            lod_min_clamp: 0.0,
-            lod_max_clamp: 0.0,
-            max_anisotropy: 0,
-            compare_function: wgpu::CompareFunction::Always,
-            border_color: wgpu::BorderColor::TransparentBlack,
-        });
-        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &bind_group_layout,
-            bindings: &[
-                wgpu::Binding {
-                    binding: 0,
-                    resource: wgpu::BindingResource::Buffer {
-                        buffer: &uniform_buf,
-                        range: 0 .. mem::size_of::<GlobalConstants>() as u32,
-                    },
-                },
-                wgpu::Binding {
-                    binding: 1,
-                    resource: wgpu::BindingResource::Sampler(&palette_sampler),
-                },
-            ],
-        });
-
-        GlobalContext {
-            uniform_buf,
-            palette_sampler,
-            bind_group_layout,
-            bind_group,
-        }
     }
 }
 
@@ -329,7 +239,7 @@ impl Palette {
 
 
 pub struct Render {
-    global: GlobalContext,
+    global: global::Context,
     object: object::Context,
     terrain: terrain::Context,
     pub light_config: settings::Light,
@@ -352,7 +262,7 @@ impl Render {
         let mut init_encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
             todo: 0,
         });
-        let global = GlobalContext::new(device);
+        let global = global::Context::new(device);
         let object = object::Context::new(&mut init_encoder, device, object_palette, &global);
         let terrain = terrain::Context::new(&mut init_encoder, device, level, &global);
         device.get_queue().submit(&[
@@ -439,7 +349,7 @@ impl Render {
 
         let mut updater = Updater::new(device);
         updater.update(&self.global.uniform_buf, &[
-            GlobalConstants::new(cam, &self.light_config),
+            global::Constants::new(cam, &self.light_config),
         ]);
         updater.update(&self.terrain.uniform_buf, &[
             terrain::Constants::new(&targets.extent)
